requirements.txt
asgiref==3.7.2
bidict==0.22.1
channels==4.0.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
enum-compat==0.0.3
eventlet==0.33.3
greenlet==3.0.1
h11==0.14.0
monotonic==1.6
netifaces==0.10.6
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
wsproto==1.2.0
